Working Towards Fully Parameterized LTSpice Model
Stated processing the LTSpice library file into a unified file. From there, the essential information is extracted and placed in a CSV with the spice model. The goal is to have the CSV data be able to be processed and the subckt data be inserted as a spice directive into the schematic to adjust the inductor and capacitor values. The process still needs a lot of refinement and adjustment before it is reasonably reliable.