Main to Renderer IPC Call with Parameters
Figured out how to make calls from main to renderer, which could be used to call into Recoil Nexus to update the state from main process data. Once an efficient way of handling this is figured out, then the serial data in main can be passed to the renderer for display easily.