First Signs of Native / Cross Toolchain Building
Using CMake presets, having two toolchains in the same CMake file is almost possible. Currently the Windows build prints hello world and the RP2040 build starts to build the main firmware (it errors but shows the correct output up until that point). This could make it much much easier to manage unit testing as invaliding the cache is not required and switching between toolchains is extremely easy.