Index Based Component Selection
Due to value based component selection returning many components which makes it harder to track through generations, now it can be index based, and with the library parts sorted by value, by moving up in index, larger component values are chosen, and vice versa. That way the evolution individuals only need to keep track of the index and mutate the index.