Simulation Bug Fixes in Schematic and Raw Processor
The schematic was using a current load but that caused the circuit to behave erratically, so it was replaced with a simple resistor. Also fixed a raw trace data processing bug where the efficiency would be negative due to LTSpice considering the current from the source to be negative. Both of these changes resulted in many more circuits successfully passing the requirements.