File Exists Checking and Better Mutating
The script now checks if the simulation it is about to run has already been ran, in which case it just uses the existing raw values instead of rerunning the simulation. Also improved the mutation function to not completely change the circuit but gradually modify it to hopefully avoid mutating into non-working circuits.