Average Input Current Discrepancy Fix
The average current calculated by the raw data and by LTSpice were wildly different by a multiple of almost 3. From research it seems like LTSpice does some data smoothing/interpolation before it calculates the average so the easiest solution is to switch to using the LTSpice .meas directives and let LT do the math for everything that is currently be calculated and just extract the values from the log file.