Functional MPPT Control
Found a bug in the MPPT implementation which returned the duty cycle multiplier inverted in the first branch of the outermost if loop. By changing the signs, the problem was fixed. Also added the logic to actually set the duty cycle to the value for the returned MPPT duty cycle.